pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-watchdog"
version = "0.1.0"
description = "Stop your AI coding assistant from being lazy. Catches scope reduction, option offering, deferral, and instruction violations in real-time."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Bhupendra Singh" },
]
keywords = ["ai", "claude", "copilot", "cursor", "watchdog", "linting", "coding-assistant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0,<7.0",
    "click>=8.0,<9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
ai-watchdog = "ai_watchdog.cli:main"

[project.urls]
Homepage = "https://github.com/bhusingh/ai-watchdog"
Repository = "https://github.com/bhusingh/ai-watchdog"
Issues = "https://github.com/bhusingh/ai-watchdog/issues"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
